wip generator cleanup
